requirements.txt
certifi==2019.3.9
chardet==3.0.4
GDAL==2.4.0
idna==2.8
Pillow==6.0.0
requests==2.21.0
requests-cache==0.5.0
urllib3==1.24.2
